delay-kfunc: Calibrate both inner and outer loop counts
To make it possible to hit the whole range of delays, we need to vary
the number of iterations in the inner loop as well, or we'll hit the
loop counter limit for the outer loop. So start with a lower number of
iterations for the inner loop, and increase it as part of the
calibration as needed.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>