netstacklat: Protect out-of-order tracking with spinlocks
Currently there is no synchronization between the updates and checks
of the per-socket ooo-range used to filter out potentially HOL-blocked
measurements. While the kernel holds a socket lock when the ooo-range
is updated in tcp_data_queue_ofo(), the kernel does not seem to hold a
socket lock by the time we check the ooo-range in
tcp_recv_timestamp(). To prevent the check of the ooo-range from
reading the ooo-range while it's concurrently updated (resulting in a
potentially inconsistent state), add a bpf_spin_lock to protect
updates and checks of the ooo-range.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>