netstacklat: Make HOL-blocking filtering configurable
To make netstacklat only report latency based on the local network and
application stack, it by default excludes TCP reads that may be
head-of-line (HOL) blocked by a missing segment.

However, in some scenarios, it may be desirable to still include these
HOL-blocked measurements. For example, in a data center where the full
network is under your control, high latency caused by drops in the
data center network could still be an issue you wish netstacklat to
detect. Therefore, add the -y/--include-tcp-hol-delay option, which
disables the mechanism used to filter out potentially HOL-blocked
reads.

An alternative design to simply disabling the filtering mechanism,
would be to separate the potentially HOL-blocked measurements into
their own histogram (e.g. "tcp-socket-read-hol-blocked"). However,
that would further increase the data cardinality (which may be a
concern when using it together ebpf-exporter). Furthermore, by simply
disabling the filtering instead we can reduce the overhead a bit and
also allow it as an option on systems were the tcp_data_queue_ofo()
function cannot be probed.

Signed-off-by: Simon Sundberg <simon.sundberg@kau.se>