nat64: Fix ICMPv6 to ICMP translation
The ICMPv6 to ICMP translation of the ICMPv6 Parameter Problem error
code swapped the two cases of the icmp6_code. Fix this by moving the
inner translation check to the right switch branch.

Also fix indentation of the switch statement to be consistent with the
rest of the function.

Fixes #134.

Signed-off-by: Toke Høiland-Jørgensen <toke@redhat.com>